Separate ExecutionPlan op construction logic into different files (#1352)
* Separate ExecutionPlan op construction logic into different files

* Address PR comments

* Refactor header files

* Address PR comments

(cherry picked from commit 4898fdedbd1bd55d9b07137c7dff6db486d2c8a4)
