+++ src/execution_plan/ops/op_apply.c
@@ -5,6 +5,7 @@
  */
 
 #include "op_apply.h"
+#include "../execution_plan_build/execution_plan_modify.h"
 
 /* Forward declarations. */
 static OpResult ApplyInit(OpBase *opBase);
